src/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
python-dotenv==0.10.3
pyyaml==5.1.1
portalocker==1.5.0
gevent==1.4.0